Make flag status an int and add ConfigManager
Convert flag status from string to int across models, database schema,
queries, protocols and tests
Add a ConfigManager singleton for server configuration and inject it
into
Runner, handlers and related components
Update sqlc config and internal package paths, add controllers module,
and misc refactors (slice append optimization, import tweaks)

+++ cookiefarm/server/internal/core/flag_ttl.go
@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-func (r *Runner) ValidateFlagTTL(ctx context.Context, flagTTL uint64, tickTime int) {
+func (r *Runner) ValidateFlagTTL(ctx context.Context, flagTTL uint64, tickTime uint64) {
 	interval := time.Duration(tickTime) * time.Second
 	if interval <= 0 {
 		logger.Log.Warn().Msgf("Invalid tickTime %d, defaulting to 60 seconds", tickTime)
@@ -15,7 +15,7 @@ func (r *Runner) ValidateFlagTTL(ctx context.Context, flagTTL uint64, tickTime i
 	ticker := time.NewTicker(interval)
 	defer ticker.Stop()
 
-	totalSecond := flagTTL * uint64(tickTime)
+	totalSecond := flagTTL * tickTime
 	for {
 		logger.Log.Debug().Msg("Waiting for clean flags process...")
 		select {
